feat : Modified the courses under Role Based Roadmaps section to remove skill  based courses , Renamed rolebasedRoadmaps.json , added and deleted a few courses from skillbasedRoadmaps , Updated Footer.jsx and Roadmap.jsx to maintain consistency in filename
