--- client/src/assets/json/roadmap.json
@@ -1,187 +1,0 @@
-[
-    {
-        "roadmap_name": "Frontend",
-        "roadmap_link": "https://roadmap.sh/frontend",
-        "icon": "🎨"
-    },
-    {
-        "roadmap_name": "Backend",
-        "roadmap_link": "https://roadmap.sh/backend",
-        "icon": "⚙️"
-    },
-    {
-        "roadmap_name": "DevOps",
-        "roadmap_link": "https://roadmap.sh/devops",
-        "icon": "🚀"
-    },
-    {
-        "roadmap_name": "Full Stack",
-        "roadmap_link": "https://roadmap.sh/full-stack",
-        "icon": "🔄"
-    },
-    {
-        "roadmap_name": "AI Engineer",
-        "roadmap_link": "https://roadmap.sh/ai-engineer",
-        "icon": "🤖"
-    },
-    {
-        "roadmap_name": "Data Analyst",
-        "roadmap_link": "https://roadmap.sh/data-analyst",
-        "icon": "📊"
-    },
-    {
-        "roadmap_name": "Software Architect",
-        "roadmap_link": "https://roadmap.sh/software-architect",
-        "icon": "🏛️"
-    },
-    {
-        "roadmap_name": "Software Design and Architecture",
-        "roadmap_link": "https://roadmap.sh/software-design-architecture",
-        "icon": "🏗️"
-    },
-    {
-        "roadmap_name": "JavaScript",
-        "roadmap_link": "https://roadmap.sh/javascript",
-        "icon": "🟨"
-    },
-    {
-        "roadmap_name": "Node.js",
-        "roadmap_link": "https://roadmap.sh/nodejs",
-        "icon": "🟢"
-    },
-    {
-        "roadmap_name": "TypeScript",
-        "roadmap_link": "https://roadmap.sh/typescript",
-        "icon": "🔷"
-    },
-    {
-        "roadmap_name": "Python",
-        "roadmap_link": "https://roadmap.sh/python",
-        "icon": "🐍"
-    },
-    {
-        "roadmap_name": "System Design",
-        "roadmap_link": "https://roadmap.sh/system-design",
-        "icon": "🔧"
-    },
-    {
-        "roadmap_name": "Java",
-        "roadmap_link": "https://roadmap.sh/java",
-        "icon": "☕"
-    },
-    {
-        "roadmap_name": "Spring Boot",
-        "roadmap_link": "https://roadmap.sh/spring-boot",
-        "icon": "🍃"
-    },
-    {
-        "roadmap_name": "PostgreSQL",
-        "roadmap_link": "https://roadmap.sh/postgresql-dba",
-        "icon": "🐘"
-    },
-    {
-        "roadmap_name": "SQL",
-        "roadmap_link": "https://roadmap.sh/sql",
-        "icon": "🗄️"
-    },
-    {
-        "roadmap_name": "Blockchain",
-        "roadmap_link": "https://roadmap.sh/blockchain",
-        "icon": "⛓️"
-    },
-    {
-        "roadmap_name": "QA",
-        "roadmap_link": "https://roadmap.sh/qa",
-        "icon": "🔍"
-    },
-    {
-        "roadmap_name": "Software Testing",
-        "roadmap_link": "https://roadmap.sh/software-testing",
-        "icon": "🧪"
-    },
-    {
-        "roadmap_name": "Cyber Security",
-        "roadmap_link": "https://roadmap.sh/cyber-security",
-        "icon": "🛡️"
-    },
-    {
-        "roadmap_name": "UX Design",
-        "roadmap_link": "https://roadmap.sh/ux-design",
-        "icon": "✨"
-    },
-    {
-        "roadmap_name": "React",
-        "roadmap_link": "https://roadmap.sh/react",
-        "icon": "⚛️"
-    },
-    {
-        "roadmap_name": "Vue",
-        "roadmap_link": "https://roadmap.sh/vue",
-        "icon": "💚"
-    },
-    {
-        "roadmap_name": "Angular",
-        "roadmap_link": "https://roadmap.sh/angular",
-        "icon": "🅰️"
-    },
-    {
-        "roadmap_name": "Design System",
-        "roadmap_link": "https://roadmap.sh/design-system",
-        "icon": "🎯"
-    },
-    {
-        "roadmap_name": "Computer Science",
-        "roadmap_link": "https://roadmap.sh/computer-science",
-        "icon": "💻"
-    },
-    {
-        "roadmap_name": "Flutter",
-        "roadmap_link": "https://roadmap.sh/flutter",
-        "icon": "🦋"
-    },
-    {
-        "roadmap_name": "Android",
-        "roadmap_link": "https://roadmap.sh/android",
-        "icon": "🤖"
-    },
-    {
-        "roadmap_name": "ASP.NET Core",
-        "roadmap_link": "https://roadmap.sh/aspnet-core",
-        "icon": "🔵"
-    },
-    {
-        "roadmap_name": "C++",
-        "roadmap_link": "https://roadmap.sh/cpp",
-        "icon": "➕"
-    },
-    {
-        "roadmap_name": "Go",
-        "roadmap_link": "https://roadmap.sh/golang",
-        "icon": "🐹"
-    },
-    {
-        "roadmap_name": "Rust",
-        "roadmap_link": "https://roadmap.sh/rust",
-        "icon": "🦀"
-    },
-    {
-        "roadmap_name": "MongoDB",
-        "roadmap_link": "https://roadmap.sh/mongodb",
-        "icon": "🍃"
-    },
-    {
-        "roadmap_name": "Docker",
-        "roadmap_link": "https://roadmap.sh/docker",
-        "icon": "🐳"
-    },
-    {
-        "roadmap_name": "Kubernetes",
-        "roadmap_link": "https://roadmap.sh/kubernetes",
-        "icon": "☸️"
-    },
-    {
-        "roadmap_name": "GraphQL",
-        "roadmap_link": "https://roadmap.sh/graphql",
-        "icon": "📈"
-    }
-]
