+++ client/src/assets/json/skillbasedRoadmaps.json
@@ -139,6 +139,13 @@
         "skill_link": "https://roadmap.sh/mongodb",
         "icon": "🍃"
     },
+
+    {
+        "roadmap_name": "UX Design",
+        "roadmap_link": "https://roadmap.sh/ux-design",
+        "icon": "✨"
+    },
+
     {
         "skill_name": "Prompt Engineering",
         "skill_link": "https://roadmap.sh/prompt-engineering",
